requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1.4
django-heroku==0.3.1
filelock==3.12.2
Flask==2.3.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
MouseInfo==0.1.3
nltk==3.8.1
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
platformdirs==3.7.0
psycopg2==2.9.6
PyAutoGUI==0.9.53
pygame==2.3.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pyscroll==2.30
python-slugify==8.0.1
PyTMX==3.31
pytweening==1.0.7
pywhatkit==5.4
regex==2023.5.5
requests==2.30.0
soupsieve==2.4.1
sqlparse==0.4.3
text-unidecode==1.3
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2022.7
urllib3==2.0.2
virtualenv==20.23.1
Werkzeug==2.3.4
whitenoise==6.5.0
wikipedia==1.4.0
